// src/app/models/pessoa-crud.model.ts
export class PessoaList {
    id: number = 0;
    nome: string = '';
    cpf: string = '';
    saldoAtual: number = 0;
    dataCadastro: Date = '' as unknown as Date;
    situacao: string = '';
    filterConcat?: string = '';
}